requirements.txt
diskcache==5.6.3
Jinja2==3.1.4
llama_cpp_python==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
Pygments==2.18.0
rich==13.9.4
typing_extensions==4.12.2
